+++ src/py/flwr/client/numpy_client.py
@@ -27,6 +27,7 @@ from flwr.common import (
     EvaluateRes,
     FitIns,
     FitRes,
+    GetParametersIns,
     GetParametersRes,
     GetPropertiesIns,
     GetPropertiesRes,
@@ -76,7 +77,7 @@ class NumPyClient(ABC):
         ----------
         config : Config
             Configuration parameters requested by the server.
-            This can be used to tell the client which parameters
+            This can be used to tell the client which properties
             are needed along with some Scalar attributes.
 
         Returns
@@ -88,8 +89,15 @@ class NumPyClient(ABC):
         """
 
     @abstractmethod
-    def get_parameters(self) -> List[np.ndarray]:
+    def get_parameters(self, config: Dict[str, Scalar]) -> List[np.ndarray]:
         """Return the current local model parameters.
+
+        Parameters
+        ----------
+        config : Config
+            Configuration parameters requested by the server.
+            This can be used to tell the client which parameters
+            are needed along with some Scalar attributes.
 
         Returns
         -------
@@ -173,15 +181,15 @@ class NumPyClientWrapper(Client):
 
     def get_properties(self, ins: GetPropertiesIns) -> GetPropertiesRes:
         """Return the current client properties."""
-        properties = self.numpy_client.get_properties(ins.config)
+        properties = self.numpy_client.get_properties(config=ins.config)
         return GetPropertiesRes(
             status=Status(code=Code.OK, message="Success"),
             properties=properties,
         )
 
-    def get_parameters(self) -> GetParametersRes:
+    def get_parameters(self, ins: GetParametersIns) -> GetParametersRes:
         """Return the current local model parameters."""
-        parameters = self.numpy_client.get_parameters()
+        parameters = self.numpy_client.get_parameters(config=ins.config)
         parameters_proto = weights_to_parameters(parameters)
         return GetParametersRes(parameters=parameters_proto)